Create README.md
This a simple real estate listing site that lists available houses in a given area and gives details such as number of rooms and price. The admin has the right to add a new listing, update and delete. The visitors to the site only see the house details. This uses django and some HTML and tailwind CSS.